Remove Dependabot auto-merge function
This workflow gives the impression that any Dependabot changes are
auto-deployed, but that's not currently the case. Rather than fixing it, we've
decided that we will manually approve all Dependabot PRs, which will then be
deployed as soon as we merge them. The ADR has more information about the
reasoning.