WIP: work around caching properties in tests
This is a little bit messy, but some tests were not actually
using the mocked HTTP responses provided in tests.

`JobServerReport`, by default, uses a `requests_cache` session
that stores HTTP responses.

This, by default, does not have a default expiry time for requests,
as `expires_after` is set to `-1`.

Some of the tests were therefore using the cached responses on
`JobServerReport(report).get_html()`, which is probably not was
intended:

* the expected value was being provided anyway, with a response that
  used it, so the intent was that the response should populate the
  resulting data value
* it would mean that the behaviour of tests could be influenced by
  the ordering of other tests (the result gets cached and then reused;
  and sharing of state this way seems like it can create or mask other
  problems)

This was masked by all the expected HTML content being the same here,
and there's an argument that perhaps the content should be the name of
the test function or something else unique.

Alternative fixes:

* name the cache and therefore have a unique cache for each test
  (however, this is currently set as a class attribute, configured by
  environment variable; it would be better if we could also specify
  it when creating the object, maybe)
* maybe make the caching less aggressive by default, and use a frozen
  time to expire the cache.