Pin `django` to 5.2 and revert, and `urllib3` update.
We decided not to do a major version upgrade of our core dependencies without
extra testing. This repository is low risk as it is not very complicated, but
that still applies.

I accidentally upgraded to Django 6.0 while doing the dependency rota. The site
is still alive at https://reports.opensafely.org/ so it doesn't appear to have
done much harm, if any. But we should still revert until we're confident.

This also includes a bugfix update to `urllib3`.